fix: move to `bin.ts`
